component renamed and added main component
