+++ src/components/guided-modal-tour/editor.scss
