+++ src/components/modal-design-library/modal.js
@@ -6,7 +6,7 @@ import BlockList from './block-list'
 import Button from '../button'
 // import AdvancedToolbarControl from '../advanced-toolbar-control'
 import DesignLibraryList from '~stackable/components/design-library-list'
-import { ModalTour } from '~stackable/components'
+import { GuidedModalTour } from '~stackable/components'
 import { getDesigns, filterDesigns } from '~stackable/design-library'
 
 /**
@@ -217,8 +217,7 @@ export const ModalDesignLibrary = props => {
 		>
 			<div className="ugb-modal-design-library__wrapper">
 
-				{ /* DEV NOTE: this is just a test */ }
-				<ModalTour />
+				<GuidedModalTour tourId="design-library-welcome" />
 
 				<aside className="ugb-modal-design-library__sidebar">
 					<div className="ugb-modal-design-library__filters">
